Add STARTING_PLAYER tile handling in give method
This commit introduces an enhancement to the `give` method of the `UsedTiles` class. It now includes a check to handle `STARTING_PLAYER` tiles differently, as per the game's rules. The `give` method will now skip adding `STARTING_PLAYER` tiles to the `usedTiles` list or perform a specified action when such tiles are encountered. A new unit test has been added to ensure that `STARTING_PLAYER` tiles are processed correctly and that normal tiles are added as expected.